+++ miso-web/src/it/resources/db/migration/plainSample_integration_test_data.sql
@@ -142,8 +142,8 @@ INSERT INTO Library(libraryId, name, alias, identificationBarcode, description, 
 INSERT INTO Library_Index(library_libraryId, index_indexId) VALUES
   (1, 5);
 
-INSERT INTO LibraryAliquot (aliquotId, name, concentration, concentrationUnits, libraryId, identificationBarcode, creationDate, creator, lastModifier, lastUpdated) VALUES
-(1, 'LDI1', 5.9, 'NANOGRAMS_PER_MICROLITRE', 1, 'LDI1::PLAIN_L0001_1-1', '2017-07-20', 1, 1, '2017-07-20 09:01:00');
+INSERT INTO LibraryAliquot (aliquotId, name, alias, concentration, concentrationUnits, libraryId, identificationBarcode, creationDate, creator, lastModifier, lastUpdated) VALUES
+(1, 'LDI1', 'PLAIN_L0001-1_1', 5.9, 'NANOGRAMS_PER_MICROLITRE', 1, 'LDI1::PLAIN_L0001_1-1', '2017-07-20', 1, 1, '2017-07-20 09:01:00');
 
 INSERT INTO Pool (poolId, concentration, volume, name, alias, identificationBarcode, description, creationDate, platformType, lastModifier, creator, created, lastModified, qcPassed) VALUES
 (1, 8.25, NULL, 'IPO1', 'POOL_1', 'IPO1::POOL_1', NULL, '2017-07-20', 'ILLUMINA', 1, 1, '2017-07-20 10:01:00', '2017-07-20 10:01:00', NULL);